+++ cmd/ore/equinixmetal/create-device.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package packet
+package equinixmetal
 
 import (
 	"encoding/json"
@@ -29,8 +29,8 @@ import (
 var (
 	cmdCreateDevice = &cobra.Command{
 		Use:   "create-device [options]",
-		Short: "Create Packet device",
-		Long:  `Create a Packet device.`,
+		Short: "Create EquinixMetal device",
+		Long:  `Create an EquinixMetal device.`,
 		RunE:  runCreateDevice,
 	}
 	hostname     string
@@ -38,13 +38,13 @@ var (
 )
 
 func init() {
-	Packet.AddCommand(cmdCreateDevice)
+	EquinixMetal.AddCommand(cmdCreateDevice)
 	cmdCreateDevice.Flags().StringVar(&options.Facility, "facility", "sjc1", "facility code")
 	cmdCreateDevice.Flags().StringVar(&options.Plan, "plan", "", "plan slug (default board-dependent, e.g. \"baremetal_0\")")
 	cmdCreateDevice.Flags().StringVar(&options.Board, "board", "amd64-usr", "Container Linux board")
 	cmdCreateDevice.Flags().StringVar(&options.InstallerImageBaseURL, "installer-image-base-url", "", "installer image base URL, non-https (default board-dependent, e.g. \"http://stable.release.flatcar-linux.net/amd64-usr/current\")")
-	cmdCreateDevice.Flags().StringVar(&options.InstallerImageKernelURL, "installer-image-kernel-url", "", "Packet installer image kernel URL, (default installer-image-base-url/flatcar_production_pxe.vmlinuz)")
-	cmdCreateDevice.Flags().StringVar(&options.InstallerImageCpioURL, "installer-image-cpio-url", "", "Packet installer image cpio URL, (default installer-image-base-url/flatcar_production_pxe_image.cpio.gz)")
+	cmdCreateDevice.Flags().StringVar(&options.InstallerImageKernelURL, "installer-image-kernel-url", "", "EquinixMetal installer image kernel URL, (default installer-image-base-url/flatcar_production_pxe.vmlinuz)")
+	cmdCreateDevice.Flags().StringVar(&options.InstallerImageCpioURL, "installer-image-cpio-url", "", "EquinixMetal installer image cpio URL, (default installer-image-base-url/flatcar_production_pxe_image.cpio.gz)")
 	cmdCreateDevice.Flags().StringVar(&options.ImageURL, "image-url", "", "image base URL (default board-dependent, e.g. \"https://alpha.release.flatcar-linux.net/amd64-usr/current/flatcar_production_packet_image.bin.bz2\")")
 	cmdCreateDevice.Flags().StringVar(&hostname, "hostname", "", "hostname to assign to device")
 	cmdCreateDevice.Flags().StringVar(&userDataPath, "userdata-file", "", "path to file containing userdata")
@@ -52,7 +52,7 @@ func init() {
 
 func runCreateDevice(cmd *cobra.Command, args []string) error {
 	if len(args) != 0 {
-		fmt.Fprintf(os.Stderr, "Unrecognized args in packet create-device cmd: %v\n", args)
+		fmt.Fprintf(os.Stderr, "Unrecognized args in equinixmetal create-device cmd: %v\n", args)
 		os.Exit(2)
 	}