+++ kola/tests/metadata/contents.go
@@ -59,10 +59,10 @@ func init() {
 	})
 
 	register.Register(&register.Test{
-		Name:        "cl.metadata.packet",
-		Run:         verifyPacket,
+		Name:        "cl.metadata.equinixmetal",
+		Run:         verifyEquinixMetal,
 		ClusterSize: 1,
-		Platforms:   []string{"packet"},
+		Platforms:   []string{"equinixmetal"},
 		UserData:    enableMetadataService,
 		Distros:     []string{"cl"},
 	})
@@ -78,7 +78,7 @@ func verifyAzure(c cluster.TestCluster) {
 	// which is required for COREOS_AZURE_IPV4_VIRTUAL to be present
 }
 
-func verifyPacket(c cluster.TestCluster) {
+func verifyEquinixMetal(c cluster.TestCluster) {
 	verify(c, "COREOS_PACKET_HOSTNAME", "COREOS_PACKET_PHONE_HOME_URL", "COREOS_PACKET_IPV4_PUBLIC_0", "COREOS_PACKET_IPV4_PRIVATE_0", "COREOS_PACKET_IPV6_PUBLIC_0")
 }